pyproject.toml
[project]
name = "azure-haymaker"
version = "0.1.0"
description = "Generate benign service telemetry for Azure Tenant simulation"
readme = "README.md"
authors = [
    { name = "Ryan Sweet", email = "rysweet@microsoft.com" }
]
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.40.0",
    "azure-identity>=1.19.0",
    "azure-mgmt-resource>=23.2.0",
    "azure-mgmt-containerinstance>=10.1.0",
    "azure-mgmt-subscription>=3.1.1",
    "azure-mgmt-authorization>=4.0.0",
    "azure-servicebus>=7.13.0",
    "azure-keyvault-secrets>=4.8.0",
    "azure-data-tables>=12.5.0",
    "azure-cosmos>=4.7.0",
    "azure-storage-blob>=12.23.0",
    "azure-monitor-ingestion>=1.0.4",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.8.0",
    "pyright>=1.1.390",
    "pre-commit>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py313"
extend-exclude = [".claude", ".claude-trace"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "PT",  # flake8-pytest-style
    "SIM", # flake8-simplify
]
ignore = [
    "E501", # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"] # allow assert in tests

[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "strict"
reportMissingTypeStubs = false
reportUnknownMemberType = false
reportUnknownArgumentType = false
reportUnknownVariableType = false
exclude = ["src/agents"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=azure_haymaker",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=0",
]
asyncio_mode = "auto"
